rename suitability table file
