+++ component/widget/__init__.py
@@ -1,4 +1,3 @@
-from .area_sum_up import *
 from .custom_aoi_dialog import *
 from .import_aoi_dialog import *
 from .export_dialog import *
