+++ component/widget/suitability_table.py
